Add files via upload
I added the start of a class to see when the card is flipped, matched, etc. I also made an array with 8 places for photos with a function that duplicates, and randomly puts them into another array. Natalyn, I am kind of at a stopping point until we all see this code, and then we can go from there.